Move `updateJsonSchema` to `createConfigService`
